requirements.txt
# Bot de telegram
python-telegram-bot
python-dotenv

# Trabajos con imágenes
opencv-python
pillow
numpy

# Reconocimiento de Tickets
easyocr
spacy
receipt-ocr
unidecode
rapidfuzz